+++ RELEASES.md
@@ -21,6 +21,7 @@ Upcoming
 - Add integer type to FixedQuerySender
 - Fix made DefaultMessageListenerContainer configurable for Jms102
 - Remove geronimo-jms_1.1_spec from runtime
+- Bugfix in "Show Security Items - Used Authentication Entries", and add sorting (no duplicates)
 
 7.0-B1
 ---